Updated for scraperlib 3.1
zimscraperlib 3.1 enforces all openZIM recommended metadata.
Due to the very flexible nature of zimit, most of it consists in supporting setting
those metadata as well as defining fallbacks.

Illustration is now handled differently. It uses the scraperlib fallback but also
converts and resizes found favicon.

One test changed as to not expect a `Source` metadata as it is not specifying one nor
a main_url.

using wabac 2.16.5

Dropping support for Python 3.6