Fixed #119: direct link hack when inside an outer frame (kiwix-serve 3.5+)
Every HTML entry gets injected with a small script that checks for SW.
This redirects the page to the home page (with back redirect) in order to install
the SW if it is not present (assuming a direct link to the HTML entry).

This hack used to check for being run from the top most window (`window.top`) which
was the case for older kiwix-serve.
Now that kiwix-serve uses its own iframe, this condition was never met.

This fixes in-frame behavior by removing the top-window check.
I suppose the check was for safety only as all tests I conducted showed no regression;
even with multiple layers of imbricated iframes