Be more tolerant to invalid css.
When tinycss2 fails to parse, we use a regex rewriter to (trying to)
rewrite url in the css.

In case of invalid css tinycss2 doesn't not failed on parsing, but
generate a ParserError token
https://doc.courtbouillon.org/tinycss2/stable/api_reference.html?highlight=parser#tinycss2.ast.ParseError

This is on serialization that exception occurs.
We even try/except all transformation to be sure.

Fix #155