Add all statics files in wheel, even if ignored in .gitignore
While some files are in .gitignore because we do not want to commit
them, we still need them in the wheel. This is typically the case of
wombat.js which is downloaded at the build stage, must be present inside
the wheel since essential to scraper but must never be commited because
we want to retrieve it at build time.