Revisit retrieve_illustration logic to give prefer best favicons
Instead of prefering to use WARC items (or prefering to download as it
was before #202), we prefer to use the most suited favicon.

If the most suited favicon is present inside the WARC, we use it
directly to avoid downloading again something. Otherwise we download it.

If both fails (icon is not in the WARC and not available for download),
we move to the next suited favicon. And so on until we have no more
icons to consider, in which case we fallback to scraperlib default
illustration.